Remove lib.rs
This prevents documentation being generated for the lib in docs.rs.
